#1896 renamed enums and moved to new files
