+++ src/frontend/src/utils/form.ts
@@ -56,3 +56,36 @@ export const startDateTester = (startDate: Date | undefined) => {
   }
   return true;
 };
+
+export enum WPFormType {
+  CREATE = 'CREATE',
+  EDIT = 'EDIT',
+  CREATEWITHCR = 'CREATEWITHCR'
+}
+
+/**
+ * Checks if the form type is a Create CR form type
+ * @param formType the given form type
+ * @returns if the form type is a Create CR Form type
+ */
+export const isCreateCr = (formType: WPFormType): boolean => {
+  return formType === WPFormType.CREATEWITHCR;
+};
+
+/**
+ * Checks if the form type is a Create form type
+ * @param formType the given form type
+ * @returns if the form type is a Create Form type
+ */
+export const isCreate = (formType: WPFormType): boolean => {
+  return formType === WPFormType.CREATE;
+};
+
+/**
+ * Checks if the form type is an edit form type
+ * @param formType the given form type
+ * @returns if the form type is an edit Form type
+ */
+export const isEdit = (formType: WPFormType): boolean => {
+  return formType === WPFormType.EDIT;
+};
