+++ src/frontend/src/utils/routes.ts
@@ -24,7 +24,7 @@ const PROJECTS_ALL = PROJECTS + '/all';
 const PROJECTS_BY_WBS = PROJECTS + `/:wbsNum`;
 const PROJECTS_NEW = PROJECTS + `/new`;
 const WORK_PACKAGE_NEW = PROJECTS + `/work-package/new`;
-const WORK_PACKAGE_CR_NEW = PROJECTS + `/work-package/newCr`;
+const WORK_PACKAGE_CR_NEW = PROJECTS + `/work-package/new-cr`;
 
 /**************** Teams Section ****************/
 const TEAMS = `/teams`;
