+++ src/shared/src/types/project-types.ts
@@ -35,12 +35,6 @@ export enum WbsElementStatus {
   Complete = 'COMPLETE'
 }
 
-export enum WPFormType {
-  Create = 'CREATE',
-  Edit = 'EDIT',
-  CrCreate = 'CrCREATE'
-}
-
 export interface Project extends WbsElement {
   summary: string;
   budget: number;
